# Optimize the murmur2 hash function